requirements.txt
# JonnyMCP 项目依赖

# MCP框架
mcp>=1.0.0

# 网络请求和HTML解析
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 图像处理
Pillow>=10.0.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 日志和配置
pyyaml>=6.0

# 测试框架
pytest>=7.4.0
pytest-cov>=4.1.0

# 开发工具
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# 类型提示
typing-extensions>=4.7.0

# 异步支持
asyncio>=3.4.3

# JSON处理
jsonschema>=4.19.0 